fix(agent): harden sql and email guards against redos
